test: move regression tests to tests/ and fix CI
Move regression tests from src/regression_tests.rs to
tests/regression_tests.rs per review feedback. Fix clippy
warnings and formatting.
